+++ packages/credential-sdk/src/keypairs/index.js
@@ -1,4 +1,4 @@
-export { default as DockKeypair } from './keypair';
+export { default as DockKeypair } from './dock-keypair';
 export { default as Ed25519Keypair } from './keypair-ed25519';
 export { default as Secp256k1Keypair } from './keypair-secp256k1';
 export { default as DidKeypair } from './did-keypair';